rename connect package to csr
