+++ internal/consul/certificates/csr/generate.go
@@ -1,4 +1,4 @@
-package connect
+package csr
 
 import (
 	"bytes"
