+++ internal/consul/certificates/csr/generate_test.go
@@ -1,4 +1,4 @@
-package connect
+package csr
 
 import (
 	"fmt"
